requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.8
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.18.26
botocore==1.21.26
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
dnspython==1.16.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fastjsonschema==2.16.2
filelock==3.0.12
Flask==2.0.1
fonttools==4.37.3
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.1
idna==2.10
importlib-metadata==4.8.1
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jira==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyring==23.2.1
kiwisolver==1.4.4
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==20.9
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.5.29
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.21.7
psutil==5.9.2
psycopg2==2.8.6
pure-eval==0.2.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.1
pytest==6.2.4
python-dateutil==2.8.2
python-decouple==3.5
python-docx==0.8.11
pytz==2021.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.12.0
Send2Trash==1.8.0
simplejwt==2.0.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
stack-data==0.5.1
tdqm==0.0.1
terminado==0.16.0
threadpoolctl==2.2.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.6
validators==0.20.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
widgetsnbextension==4.0.3
zipp==3.6.0
